Factoring out some common form mappings (schedule, data set id)
