+++ web/app/org/ada/web/controllers/dataset/datatrans/DataSetMetaTransformationFormViews.scala
@@ -5,6 +5,7 @@ import java.util.Date
 import org.incal.core.dataaccess.StreamSpec
 import org.ada.server.models._
 import org.ada.server.models.datatrans.{DataSetMetaTransformation, ResultDataSetSpec}
+import org.ada.web.controllers.MappingHelper
 import org.ada.web.controllers.core.GenericMapping
 import org.incal.core.util.toHumanReadableCamel
 import org.incal.core.util.hasNonAlphanumericUnderscore
@@ -21,7 +22,7 @@ import scala.reflect.runtime.universe.{TypeTag, typeOf}
 
 abstract protected[controllers] class DataSetMetaTransformationFormViews[E <: DataSetMetaTransformation: TypeTag](
   implicit manifest: Manifest[E]
-) extends CreateEditFormViews[E, BSONObjectID] {
+) extends CreateEditFormViews[E, BSONObjectID] with MappingHelper {
 
   private val domainNameSuffix = "Transformation"
   private val humanReadableSuffix = toHumanReadableCamel(domainNameSuffix)
@@ -32,26 +33,6 @@ abstract protected[controllers] class DataSetMetaTransformationFormViews[E <: Da
   private implicit val mapFormatter = MapJsonFormatter.apply
   private implicit val filterShowFieldStyleFormatter = EnumFormatter(FilterShowFieldStyle)
   private implicit val widgetGenerationMethodFormatter = EnumFormatter(WidgetGenerationMethod)
-  private implicit val weekDayFormatter = EnumFormatter(WeekDay)
-
-  // Basic Forms
-
-  protected val scheduledTimeMapping: Mapping[ScheduledTime] = mapping(
-    "weekDay" -> optional(of[WeekDay.Value]),
-    "hour" -> optional(number(min = 0, max = 23)),
-    "minute" -> optional(number(min = 0, max = 59)),
-    "second" -> optional(number(min = 0, max = 59))
-  )(ScheduledTime.apply)(ScheduledTime.unapply)
-
-  private val upperCasePattern = "[A-Z]".r
-
-  protected val dataSetIdMapping = nonEmptyText.verifying(
-    "Data Set Id must not contain any non-alphanumeric characters (except underscore)",
-    dataSetId => !hasNonAlphanumericUnderscore(dataSetId.replaceFirst("\\.",""))
-  ).verifying(
-    "Data Set Id must not contain any upper case letters",
-    dataSetId => !upperCasePattern.findFirstIn(dataSetId).isDefined
-  )
 
   protected val extraMappings: Traversable[(String, Mapping[_])] = Nil
